Update star-rating.js
add a new Parameter addCaptionAsTitle was added, which is set in displayOnly Mode to true but only when it isn't manually set.
So it is still possible to disable it.  

When addCaptionAsTitle  = true the caption "content" is set as the title html attribute to the stars containing div.
So you would get the correct value on hover.

The parameter showCaption isn't disabled in displayOnly  = true automatically any more. 
So the caption could be rendered in displayOnly  mode.
If the caption shouldn't be shown it has to be disabled manually.

I just get the caption content by using the text() function.
But of course an additional function which only returns the caption text like fetchCaptionValue would also be possible and maybe even better.